requirements.txt
tensorflow
pyyaml
pandas
sympy